app/bank/requirements.txt
Django==4.2.4
pytest-django==4.5.2
